requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
django-filter==24.3
django-cors-headers==4.4.0
drf-nested-routers==0.94.1
Pillow==10.4.0
sqlparse==0.5.3
tzdata==2025.2